#57 refactor: 거래 생성 에러 처리
